Exercise 5.4.2-2 (Difficult) - extend shapeBounds
This was a real pain. I took the instructions to mean that the bounds
could be no larger than the clipped rect, but that they could be
smaller. I fussed about endlessly with Math.min/max, but couldn't solve
the logic of it well. Finally I remembered there were bounds boolean
functions in the file, so I just did an intersection. Then the problem
I'd been having before reappeared, namely that things go wacky when the
picture is entirely outside the clipping region. I labored over the idea
some more, and finally gave up, and decided to just set the clipped
rect as the new bounds rect. The easiest way to get these bounds was to
reuse the bounds of a line between the given points.